move user models into its own dir, update references
